+++ DocuSignSDK.xcframework/ios-arm64_armv7/DocuSignSDK.framework/Headers/DSMEnvelopeDefinition.h
@@ -1,0 +1,8 @@
+#import <Foundation/Foundation.h>
+#import <DocuSignSDK/DSMEnvelopeTemplate.h>
+
+@interface DSMEnvelopeDefinition : DSMEnvelopeTemplate
+
+@property (nonatomic, strong) NSString *envelopeName;
+
+@end