+++ DocuSignSDK.xcframework/ios-arm64_armv7/DocuSignSDK.framework/Headers/DSMEnvelopeTextCustomField.h
@@ -1,0 +1,8 @@
+#import <DocuSignSDK/DSMTextCustomField.h>
+
+@interface DSMEnvelopeTextCustomField : DSMTextCustomField
+
+- (instancetype) init NS_UNAVAILABLE;
+
+@end
+