+++ DocuSignSDK.xcframework/ios-arm64_i386_x86_64-simulator/DocuSignSDK.framework/Headers/DSMEnvelopeDocument.h
@@ -1,0 +1,9 @@
+#import <DocuSignSDK/DSMDocument.h>
+
+@interface DSMEnvelopeDocument : DSMDocument
+
+- (instancetype)init NS_UNAVAILABLE;
+
+@end
+
+