+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # DocuSign Native iOS SDK Changelog
 
+## [v2.5.1] - 06/14/2021
+
+### Added
+* Native-iOS-SDK is built with Xcode12.5 and now includes XCFramework as the primary framework for CocoaPods distribution.
+
+### Fixed
+* Online signing for failing envelopes returns signing-cancelled notification with `error` details in `userInfo`.
+
 ## [v2.5] - 05/12/2021
 
 ### Deprecated